+++ UVOCBot.Api/Startup.cs
@@ -4,6 +4,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.OpenApi.Models;
+using UVOCBot.Api.Workers;
 
 namespace UVOCBot.Api
 {
@@ -23,6 +24,7 @@ namespace UVOCBot.Api
             services.AddDbContext<BotContext>();
             services.AddControllers();
             services.AddSwaggerGen(c => c.SwaggerDoc("v1", new OpenApiInfo { Title = "UVOCBot.Api", Version = "v1" }));
+            services.AddHostedService<CleanupWorker>();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.