+++ UVOCBot/Commands/TeamGenerationModule.cs
@@ -8,13 +8,12 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using UVOCBot.Exceptions;
-using UVOCBot.Extensions;
 
 namespace UVOCBot.Commands
 {
     [Description("Commands that help with generating team lists")]
     [RequireGuild]
-    public class TeamsModule : BaseCommandModule
+    public class TeamGenerationModule : BaseCommandModule
     {
         [Command("random-teams")]
         [Aliases("rt", "random")]