+++ UVOCBot/Extensions/MemberReturnedInfo.cs
@@ -1,13 +1,13 @@
 ﻿using DSharpPlus.Entities;
 
-namespace UVOCBot.Utils
+namespace UVOCBot.Extensions
 {
     public struct MemberReturnedInfo
     {
         public enum GetMemberStatus
         {
             /// <summary>
-            /// The guild member was successfully received
+            /// The guild member was successfully retrieved
             /// </summary>
             Success = 0,
 
@@ -17,7 +17,7 @@ namespace UVOCBot.Utils
             GuildNotFound = 1,
 
             /// <summary>
-            /// Fetching the channel failed
+            /// The guild member could not be retrieved
             /// </summary>
             Failure = 2
         }